requirements.txt
evaluate~=0.4.1
langchain==0.1.19
langchain_community==0.0.38
langchain_google_community==1.0.3
langchain_core==0.1.52
langchain_google_community==1.0.3
langchain_openai==0.1.6
pandas~=2.1.1
rouge_score
torch~=2.3.0
tqdm~=4.65.0
transformers~=4.40.2
vllm~=0.4.2

regex~=2023.10.3
requests~=2.31.0
dataclasses~=0.6
datasets~=2.18.0
Deprecated~=1.2.14
matplotlib~=3.7.2
numpy~=1.24.3
seaborn~=0.13.0
google_books_api_wrapper
